pyproject.toml
[project]
name = "reactor-enterprise-integration-agents"
version = "0.1.0"
description = "Enterprise Integration Patterns with Azure AI Foundry Agents"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-projects>=1.0.0",
    "azure-identity>=1.19.0",
    "azure-eventhub>=5.11.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agents", "src/shared", "src/patterns"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint.isort]
known-first-party = ["agents", "shared", "patterns"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
mypy_path = "src"

[tool.pytest.ini_options]
pythonpath = ["src"]
